moving home files to more specific package
