+++ app/src/main/java/br/com/rbrthmn/ui/financialcompanion/screens/home/components/balancecard/BalanceCard.kt
@@ -18,7 +18,7 @@
  *
  */
 
-package br.com.rbrthmn.ui.financialcompanion.screens.home.components
+package br.com.rbrthmn.ui.financialcompanion.screens.home.components.balancecard
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
@@ -52,6 +52,9 @@ import androidx.compose.ui.unit.sp
 import androidx.compose.ui.window.Dialog
 import br.com.rbrthmn.R
 import br.com.rbrthmn.ui.financialcompanion.common.DecimalInputField
+import br.com.rbrthmn.ui.financialcompanion.screens.home.components.AddItemButton
+import br.com.rbrthmn.ui.financialcompanion.screens.home.components.BanksDropdownMenu
+import br.com.rbrthmn.ui.financialcompanion.screens.home.components.TotalValueText
 import org.koin.androidx.compose.koinViewModel
 
 @Composable
