+++ app/src/main/java/br/com/rbrthmn/ui/financialcompanion/screens/home/components/balancecard/BankAccountBalanceUiState.kt
@@ -18,9 +18,14 @@
  *
  */
 
-package br.com.rbrthmn.ui.financialcompanion.screens.home.components
+package br.com.rbrthmn.ui.financialcompanion.screens.home.components.balancecard
 
-data class BalanceCardUiState(
-    val totalBalance: String = "",
-    val accounts: List<BankAccountBalanceUiState> = listOf()
+import br.com.rbrthmn.R
+
+data class BankAccountBalanceUiState(
+    val name: String = "",
+    val value: String = "",
+    val bankName: String = "",
+    val bankIcon: Int = R.drawable.bank_icon,
+    val canValueBeEdited: Boolean = false
 )
