+++ app/src/main/java/br/com/rbrthmn/ui/financialcompanion/screens/home/components/creditcardbillscard/CreditCardBillUiState.kt
@@ -18,7 +18,7 @@
  *
  */
 
-package br.com.rbrthmn.ui.financialcompanion.screens.home.components
+package br.com.rbrthmn.ui.financialcompanion.screens.home.components.creditcardbillscard
 
 import br.com.rbrthmn.R
