+++ app/src/main/java/br/com/rbrthmn/ui/financialcompanion/screens/home/components/creditcardbillscard/CreditCardsBillCardUiState.kt
@@ -18,14 +18,9 @@
  *
  */
 
-package br.com.rbrthmn.ui.financialcompanion.screens.home.components
+package br.com.rbrthmn.ui.financialcompanion.screens.home.components.creditcardbillscard
 
-import br.com.rbrthmn.R
-
-data class BankAccountBalanceUiState(
-    val name: String = "",
-    val value: String = "",
-    val bankName: String = "",
-    val bankIcon: Int = R.drawable.bank_icon,
-    val canValueBeEdited: Boolean = false
+data class CreditCardsBillCardUiState(
+    val totalBill: String = "",
+    val bills: List<CreditCardBillUiState> = listOf()
 )
